runtime/internal/sys: remove 386 assembly code
Reviewed-on: https://go-review.googlesource.com/c/go/+/475736
Run-TryBot: Wayne Zuo <wdvxdr@golangcn.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
